+++ packages/core/src/test.ts
@@ -2,7 +2,8 @@
 
 import { ThemeRegistry } from '@aesthetic/system';
 import { darkTheme, design, lightTheme } from '@aesthetic/system/test';
-import { ThemeName } from '@aesthetic/types';
+import { ClassName, Engine, RenderOptions, RenderResult, Rule, ThemeName } from '@aesthetic/types';
+import { hyphenate, objectLoop } from '@aesthetic/utils';
 import { Aesthetic, AestheticOptions, EventListener, EventType, ThemeSheet } from '.';
 
 export { darkTheme, design, lightTheme };
@@ -62,4 +63,49 @@ export function teardownAesthetic(aesthetic: Aesthetic<any, any>) {
 	lightTheme.name = '';
 	darkTheme.name = '';
 	resetAestheticState(aesthetic);
+}
+
+export function createTestEngine(): Engine<Rule, ClassName> {
+	const noop = () => {};
+
+	const renderVariable = (name: string) => `variable:${hyphenate(name)}`;
+
+	const renderRule = (rule: Rule, options: RenderOptions = {}): RenderResult<ClassName> => {
+		let result = String(options.debugName ?? 'class');
+
+		objectLoop(rule['@variables'], (value, name) => {
+			result += ` ${renderVariable(name)}`;
+		});
+
+		return {
+			result,
+			variants: Object.keys(rule['@variants'] ?? {}).map((variant) => {
+				const types = variant.split('+').map((v) => v.trim());
+
+				return {
+					result: types.map((v) => `variant:${hyphenate(v)}`).join(' '),
+					types,
+				};
+			}),
+		};
+	};
+
+	const engine: Engine<Rule, ClassName> = {
+		direction: 'ltr',
+		prefersColorScheme: () => false,
+		prefersContrastLevel: () => false,
+		renderDeclaration: (prop) => `property:${hyphenate(prop)}`,
+		renderFontFace: () => 'font-face',
+		renderImport: () => 'import',
+		renderKeyframes: (keyframes, name) => `keyframes:${hyphenate(name ?? 'unknown')}`,
+		renderRule,
+		renderRuleGrouped: renderRule,
+		renderVariable,
+		ruleCount: 0,
+		setDirection: noop,
+		setRootVariables: noop,
+		setTheme: noop,
+	};
+
+	return engine;
 }
