+++ packages/style/src/test.ts
@@ -40,51 +40,6 @@ export function createTestStyleEngine(
 	});
 }
 
-export function createTestEngine(): Engine<Rule, ClassName> {
-	const noop = () => {};
-
-	const renderVariable = (name: string) => `variable:${hyphenate(name)}`;
-
-	const renderRule = (rule: Rule, options: RenderOptions = {}): RenderResult<ClassName> => {
-		let result = String(options.debugName ?? 'class');
-
-		objectLoop(rule['@variables'], (value, name) => {
-			result += ` ${renderVariable(name)}`;
-		});
-
-		return {
-			result,
-			variants: Object.keys(rule['@variants'] ?? {}).map((variant) => {
-				const types = variant.split('+').map((v) => v.trim());
-
-				return {
-					result: types.map((v) => `variant:${hyphenate(v)}`).join(' '),
-					types,
-				};
-			}),
-		};
-	};
-
-	const engine: Engine<Rule, ClassName> = {
-		direction: 'ltr',
-		prefersColorScheme: () => false,
-		prefersContrastLevel: () => false,
-		renderDeclaration: (prop) => `property:${hyphenate(prop)}`,
-		renderFontFace: () => 'font-face',
-		renderImport: () => 'import',
-		renderKeyframes: (keyframes, name) => `keyframes:${hyphenate(name ?? 'unknown')}`,
-		renderRule,
-		renderRuleGrouped: renderRule,
-		renderVariable,
-		ruleCount: 0,
-		setDirection: noop,
-		setRootVariables: noop,
-		setTheme: noop,
-	};
-
-	return engine;
-}
-
 export function getRenderedStyles(type: Sheet | SheetType): CSS {
 	return [...(typeof type === 'string' ? getStyleElement(type) : type).cssRules].reduce(
 		(css, rule) => css + rule.cssText,
